Throw together the mixin logger gem
