+++ Ruby/XasLog/README.md
@@ -1,0 +1,9 @@
+
+# XasLogger - a colorized, multi-output mixing
+
+Really, this is just a fancy wrapper for the standard Ruby logger. I didn't like
+how it looks, I prefer the colorized output of the ESP-IDF's logger - can you blame me?
+
+Either way, this simple and smol gem provides a few helper functions and mix-ins
+to allow for multi-output logging, per-module log level setting, colorized output
+and a mix-in for extra easy logging of class based code :>
